fix: Remove Trivy installation script (#9) (#11)
As Trivy checks out its installation script into the same directory as
the workflow users own code is typically checked out into this can cause
other linters/checks that a workflow is using to fail due to the
unexpected install script being present.  This commit modifies the
action to ensure that the installation script is removed after the
action has installed trivy.